+++ frontend/src/pages/user.tsx
@@ -1,4 +1,4 @@
-import { UserTargetPermissionsOnResourceTypes } from "@components/config/util";
+import { UserTargetPermissionsOnResourceTypes } from "@components/users/resource-type-permissions";
 import { KeysTable } from "@components/keys/table";
 import { Page, Section } from "@components/layouts";
 import { PermissionsTable } from "@components/users/permissions-table";
@@ -11,8 +11,9 @@ import { useInvalidate, useRead, useWrite } from "@lib/hooks";
 import { Label } from "@ui/label";
 import { Switch } from "@ui/switch";
 import { useToast } from "@ui/use-toast";
-import { Key, UserCheck, UserMinus } from "lucide-react";
-import { useParams } from "react-router-dom";
+import { Key, UserCheck, UserMinus, Users } from "lucide-react";
+import { Link, useParams } from "react-router-dom";
+import { Button } from "@ui/button";
 
 export const UserPage = () => {
   const { toast } = useToast();
@@ -83,6 +84,7 @@ export const UserPage = () => {
       {user.config.type === "Service" && <ApiKeysTable user_id={user_id} />}
       {user.enabled && !user.admin && (
         <>
+          <Groups user_id={user_id} />
           <UserTargetPermissionsOnResourceTypes
             user_target={{ type: "User", id: user._id?.$oid! }}
           />
@@ -102,6 +104,29 @@ const ApiKeysTable = ({ user_id }: { user_id: string }) => {
       actions={<CreateKeyForServiceUser user_id={user_id} />}
     >
       <KeysTable keys={keys ?? []} DeleteKey={DeleteKeyForServiceUser} />
+    </Section>
+  );
+};
+
+const Groups = ({ user_id }: { user_id: string }) => {
+  const groups = useRead("ListUserGroups", {}).data?.filter((group) =>
+    group.users.includes(user_id)
+  );
+  if (!groups || groups.length === 0) {
+    return null;
+  }
+  return (
+    <Section title="Groups">
+      <div className="grid gap-4 grid-cols-1 md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4">
+        {groups.map((group) => (
+          <Link to={`/user-groups/${group._id?.$oid}`}>
+            <Button variant="link" className="flex gap-2 items-center p-0">
+              <Users className="w-4 h-4" />
+              {group.name}
+            </Button>
+          </Link>
+        ))}
+      </div>
     </Section>
   );
 };
